+++ tools/file_packager.py
@@ -478,9 +478,7 @@ if has_preloaded:
       Module['locateFile'] = Module['locateFilePackage'];
       err('warning: you defined Module.locateFilePackage, that has been renamed to Module.locateFile (using your locateFilePackage for now)');
     }
-    var REMOTE_PACKAGE_NAME = typeof Module['locateFile'] === 'function' ?
-                              Module['locateFile'](REMOTE_PACKAGE_BASE) :
-                              ((Module['filePackagePrefixURL'] || '') + REMOTE_PACKAGE_BASE);
+    var REMOTE_PACKAGE_NAME = Module['locateFile'](REMOTE_PACKAGE_BASE);
   ''' % (shared.JS.escape_for_js_string(data_target), shared.JS.escape_for_js_string(remote_package_name))
   metadata['remote_package_size'] = remote_package_size
   metadata['package_uuid'] = str(package_uuid)
@@ -735,9 +733,7 @@ ret += '''%s
 
  function runMetaWithFS() {
   Module['addRunDependency']('%(metadata_file)s');
-  var REMOTE_METADATA_NAME = typeof Module['locateFile'] === 'function' ?
-                             Module['locateFile']('%(metadata_file)s') :
-                             ((Module['filePackagePrefixURL'] || '') + '%(metadata_file)s');
+  var REMOTE_METADATA_NAME = Module['locateFile']('%(metadata_file)s');
   var xhr = new XMLHttpRequest();
   xhr.onreadystatechange = function() {
    if (xhr.readyState === 4 && xhr.status === 200) {